Rename RoughlyDrawableMaker to RoughlyDrawableMakable
