+++ roughfeel/examples/heart_svg_path.rs
@@ -12,7 +12,7 @@ use roughfeel::graphics::drawable_ops::OpSet;
 use roughfeel::renderer_engine::kurbo_drawable::{KurboDrawable, KurboOpSet};
 use roughfeel::renderer_engine::kurbo_drawable_maker::KurboDrawableMaker;
 
-use roughfeel::graphics::drawable_maker::{Generator, RoughlyDrawableMaker};
+use roughfeel::graphics::drawable_maker::{Generator, traits::RoughlyDrawableMakable};
 use roughfeel::graphics::paint::FillStyle;
 use roughfeel::*;
